+++ dwjb-api/extensions/users-permissions/config/jwt.js
@@ -1,0 +1,3 @@
+module.exports = {
+  jwtSecret: process.env.JWT_SECRET || '38e2de37-08e0-459b-baa6-4290e9dd3478'
+};