+++ src/components/portfolio/Projects/Projects.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Story, Meta } from '@storybook/react/types-6-0'
+import { Projects, ProjectsProps } from '.'
+
+export default {
+  title: 'Component/Projects',
+  component: Projects,
+  argTypes: {
+    title: { control: 'string' },
+  },
+} as Meta
+
+const Template: Story<ProjectsProps> = (args) => <Projects {...args} />
+
+// Default scenario
+export const Default = Template.bind({})
+Default.args = {}