+++ src/components/portfolio/Skills/index.tsx
@@ -11,90 +11,115 @@ import { ReactLogo } from '@styled-icons/boxicons-logos/ReactLogo'
 import { Mongodb } from '@styled-icons/simple-icons/Mongodb'
 import { Storybook } from '@styled-icons/simple-icons/Storybook'
 import { Vuejs } from '@styled-icons/boxicons-logos/Vuejs'
-import { Graphql } from '@styled-icons/simple-icons/Graphql'
+import { Sketch } from '@styled-icons/simple-icons/Sketch'
 import ReactTooltip from 'react-tooltip'
 
-export interface HeroProps {
+export interface SkillsProps {
   title?: string
 }
 
-export const Skills: React.FC<HeroProps> = (
-  { title } = { title: 'A simple Tailiwind template' }
-) => {
+export const Skills: React.FC<SkillsProps> = () => {
+  const dialogStyles = {
+    top: '50%',
+  }
+
   return (
     <>
-      <motion.div className="h-screen w-full border- border-green-200 container">
-        <h3 className="font-semibold font-sans text-black">Skills</h3>
-        <svg height="10" width="100" className="my-2">
-          <line
-            x1="0"
-            y1="0"
-            x2="100"
-            y2="0"
-            stroke="hsl(287, 47%, 67%)"
-            strokeWidth="6"
-          />
-        </svg>
-        <h1 className="text-2xl text-black">Skills Subheader</h1>
-        <div className="flex flex-col mx-auto border- border-red-300">
-          <div className="flex flex-row border- mt-20 border-red-200">
-            <div className="w-1/3 text-center border- border-red-100">
-              <Typescript width="64" height="64" data-tip="Typescript" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Html5 width="64" height="64" data-tip="HTML" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Csswizardry width="64" height="64" data-tip="CSS" />
-              <ReactTooltip />
+      <div className="bg-purp py-10 my-10">
+        <motion.div className="w-full border- p-4 my-30 py-10  border-green-200 container ">
+          <h1 className="text-2xl text-purp-dark payton-one">
+            Technologies &amp; tools I like to work with 👾
+          </h1>
+          <div className="items-center justify-center w-full">
+            <div className="grid grid-cols-2 md:grid-cols-3 gap-4 mt-20 mx-auto">
+              <div className="mx-auto">
+                <Typescript
+                  width="64"
+                  height="64"
+                  data-tip="Typescript"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Html5 width="64" height="64" data-tip="HTML" fill="white" />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Csswizardry
+                  width="64"
+                  height="64"
+                  data-tip="CSS"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <FileJson width="64" height="64" data-tip="JSON" fill="white" />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Mongodb
+                  width="64"
+                  height="64"
+                  data-tip="Mongo DB"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Nodejs width="64" height="64" data-tip="NodeJS" fill="white" />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Angular
+                  width="64"
+                  height="64"
+                  data-tip="Angular"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <ReactLogo
+                  width="64"
+                  height="64"
+                  data-tip="React"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Vuejs width="64" height="64" data-tip="Vue" fill="white" />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Storybook
+                  width="64"
+                  height="64"
+                  data-tip="Storybook"
+                  fill="white"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Framer
+                  width="64"
+                  height="64"
+                  data-tip="Framer"
+                  fill="white"
+                  stroke="#E0E0E0"
+                />
+                <ReactTooltip />
+              </div>
+              <div className="mx-auto">
+                <Sketch width="64" height="64" data-tip="Sketch" fill="white" />
+                <ReactTooltip />
+              </div>
             </div>
           </div>
-          <div className="flex flex-row border- mt-20 border-red-200">
-            <div className="w-1/3 text-center border- border-red-100">
-              <FileJson width="64" height="64" data-tip="JSON" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Mongodb width="64" height="64" data-tip="Mongo DB" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Nodejs width="64" height="64" data-tip="NodeJS" />
-              <ReactTooltip />
-            </div>
-          </div>
-          <div className="flex flex-row border- mt-20 border-red-200">
-            <div className="w-1/3 text-center border- border-red-100">
-              <Angular width="64" height="64" data-tip="Angular" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <ReactLogo width="64" height="64" data-tip="React" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Vuejs width="64" height="64" data-tip="Vue" />
-              <ReactTooltip />
-            </div>
-          </div>
-          <div className="flex flex-row border- mt-20 border-red-200">
-            <div className="w-1/3 text-center border- border-red-100">
-              <Storybook width="64" height="64" data-tip="Storybook" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Framer width="64" height="64" data-tip="Framer" />
-              <ReactTooltip />
-            </div>
-            <div className="w-1/3 text-center border- border-red-100">
-              <Graphql width="64" height="64" data-tip="Graph QL" />
-              <ReactTooltip />
-            </div>
-          </div>
-        </div>
-      </motion.div>
+        </motion.div>
+      </div>
     </>
   )
 }