+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## [0.0.10] - 20-08-21
+### Changes recommended by Gary Simon
+- contrast of text
+- wave svg (still not 100% responsive)
+
+
+## [0.0.9] - 04-08-21
+### Added
+- GTM
+
+## [0.0.8] - 03-08-21
+### Changed
+- put typewriter into a component
+
+## [0.0.7] - 27-07-21
+### Added
+- responsive hooks to conditionally include components based off of media queries. see src/hooks/responsive
+
+### Fixed
+- viewport issues on mobile
+
+### Changed
+- copy
+
+## [0.0.6] - 25-07-21
+### Added
+- description to positions
+- more intro text
+- GA tag
+
+### Fixed
+- z-index on linkedin icon was behind the graphic
+
+## [0.0.5] - 24-07-21
+### Added
+- venn diagram
+- headshot and intro blurb
+- better skills area and footer with link to linkedin and email address
+
+## [0.0.4] - 24-07-21
+### Added
+- Timeline components
+- Image wave component to hero
+- Strapi to Heroku - deploy via git cli from dwjb-api folder
+
+- Getting to grips more with react and adding framer-motion to components - FUN!
+
 ## [0.0.3] - 18-07-21
 ### Added
 - component for each section of home page